# backend/swagger.yaml
openapi: 3.0.0
info:
  title: SuperMook API
  description: API для приложения тренировочных собеседований
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Локальный сервер разработки
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Session:
      type: object
      properties:
        id:
          type: string
          description: Уникальный идентификатор сессии
        videoLink:
          type: string
          description: Ссылка на видеозвонок
        startTime:
          type: string
          format: date-time
          description: Время начала сессии
        status:
          type: string
          enum: [pending, active, completed]
          description: Статус сессии
        interviewerId:
          type: string
          description: ID пользователя в роли интервьюера
        intervieweeId:
          type: string
          description: ID пользователя в роли отвечающего
        observerIds:
          type: array
          items:
            type: string
          description: Массив ID пользователей в роли наблюдателей
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: Сообщение об ошибке
        details:
          type: string
          description: Детали ошибки (опционально)
security:
  - bearerAuth: []
paths:
...
  /sessions:
    get:
      summary: Получение списка всех сессий
      description: Возвращает список всех сессий в системе
      responses:
        '200':
          description: Успешное получение списка сессий
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
        '500':
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Создание новой сессии
      description: Создает новую сессию в системе
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                videoLink:
                  type: string
                  description: Ссылка на видеозвонок (опционально)
                startTime:
                  type: string
                  format: date-time
                  description: Время начала сессии (опционально, по умолчанию - текущее время)
      responses:
        '201':
          description: Успешное создание сессии
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '500':
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...